refactor: rename utils as kebab case
