+++ src/composables/useArchive.ts
@@ -5,7 +5,7 @@ import { usePopups } from '@/composables/usePopups'
 import { getElement, getElements, getDoc } from '@/utils/commons'
 import { Logger } from '@/utils/logger'
 import { ArchiveDownloadMethod } from '@/constants/monkey'
-import { quickArchiveDownloadMethod } from '@/utils/GMVariables'
+import { quickArchiveDownloadMethod } from '@/utils/gm-variables'
 import { setAsDownloaded } from '@/utils/highlight-galleries'
 
 export function useArchive() {
