+++ src/utils/e-hentai-api.ts
