+++ src/utils/gm-variables.ts
